requirements.txt
langgraph>=0.0.15
langchain-neo4j>=0.1.0
neo4j>=5.14.0
pydantic-ai>=0.0.1
streamlit>=1.31.0
python-dotenv>=1.0.0
langchain-google-genai>=1.0.1
